// src/main/frontend/src/api/models/CustomerDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { ApparelOrderDto } from './ApparelOrderDto';
/**
 * Data Transfer Object for Customer entities
 */
export type CustomerDto = {
    /**
     * Unique identifier for the customer. Read-only field generated by the system.
     */
    readonly id?: number;
    /**
     * Version number for optimistic locking. Read-only field incremented on each update.
     */
    readonly version?: number;
    /**
     * Date and time when the customer was created. Read-only field.
     */
    readonly createdDate?: string;
    /**
     * Date and time when the customer was last updated. Read-only field.
     */
    readonly updateDate?: string;
    /**
     * Name of the customer.
     */
    name: string;
    /**
     * Email address of the customer.
     */
    email?: string;
    /**
     * Phone number of the customer.
     */
    phoneNumber?: string;
    /**
     * First line of the customer's address.
     */
    addressLine1: string;
    /**
     * Second line of the customer's address (optional).
     */
    addressLine2?: string;
    /**
     * City of the customer's address.
     */
    city: string;
    /**
     * State or province of the customer's address.
     */
    state: string;
    /**
     * Postal code of the customer's address.
     */
    postalCode: string;
    /**
     * Apparel orders placed by this customer.
     */
    readonly apparelOrders?: Array<ApparelOrderDto>;
};